+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.service.film;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class FilmService {
+
+}
