+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserService.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.service.user;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+
+}
