+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.filmorate.storage.film;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.List;
+
+public interface FilmStorage {
+    Film addFilm(Film film);
+
+    Film updateFilm(Film film);
+
+    List<Film> getFilms();
+}
